+++ bindings/ergo-lib-wasm/src/transaction.rs
@@ -99,9 +99,9 @@ impl Transaction {
             .map_err(|e| JsValue::from_str(&format!("{}", e)))
     }
 
-    /// JSON representation (same as [`Self::to_json`],
-    /// but with box value and token amount encoding as strings)
-    pub fn to_json_dapp(&self) -> Result<JsValue, JsValue> {
+    /// JSON representation according to EIP-12 https://github.com/ergoplatform/eips/pull/23
+    /// (similar to [`Self::to_json`], but with box value and token amount encoding as strings)
+    pub fn to_js_eip12(&self) -> Result<JsValue, JsValue> {
         let tx_dapp: TransactionJsonDapp = self.0.clone().into();
         JsValue::from_serde(&tx_dapp).map_err(|e| JsValue::from_str(&format!("{}", e)))
     }
@@ -182,9 +182,9 @@ impl UnsignedTransaction {
             .map_err(|e| JsValue::from_str(&format!("{}", e)))
     }
 
-    /// JSON representation (same as [`Self::to_json`],
-    /// but with box value and token amount encoding as strings)
-    pub fn to_json_dapp(&self) -> Result<JsValue, JsValue> {
+    /// JSON representation according to EIP-12 https://github.com/ergoplatform/eips/pull/23
+    /// (similar to [`Self::to_json`], but with box value and token amount encoding as strings)
+    pub fn to_js_eip12(&self) -> Result<JsValue, JsValue> {
         let tx_dapp: UnsignedTransactionJsonDapp = self.0.clone().into();
         JsValue::from_serde(&tx_dapp).map_err(|e| JsValue::from_str(&format!("{}", e)))
     }